test(int-759): add command to generate test coverage files update vitest config create a new page structure to handle unit tests change richTestResolver test to use the file richTestResolver insteadof the class instance
